+++ pambox/speech/experiment.py
@@ -490,6 +490,41 @@ class Experiment(object):
         print(dsrts)
         return srts, dsrts
 
+    def srts_from_df(self, df, col='Intelligibility', srt_at=50):
+        """Get dataframe with SRTs
+
+        Parameters
+        ----------
+        df : Data Frame
+            DataFrame resulting from an experiment. It must have an
+            "Intelligibility" column.
+        col : string (optional)
+            Name of the column to use for the SRT calculation. The default
+            value is the 'Intelligibility' column.
+        srt_at : float (optional)
+            Value corresponding to the SRT. The default is 50 (%). This
+            parameter can be used to calculate SRT directly from other
+            metric, such as the SII. Simply set `srt_at` to the SII criterion.
+
+        Returns
+        -------
+        out : Data frame
+            Data frame, with an SRT column.
+        """
+        groups = self._get_groups(df)
+        grouped = df.groupby(groups).mean()
+        # Get name of all groups, which was now the index.
+        # Remove the SNR column because it should not be a group
+        # when calculating the SRTs.
+        grp_for_int = list(set(grouped.index.names) - {'SNR'})
+        # Get the SNR values for the tranformation in intelligibility.
+        snrs = df['SNR'].unique()
+        fc_to_srt = lambda y: int2srt(snrs, y, srt_at=srt_at)
+
+        grouped = grouped.reset_index()
+        return grouped.groupby(grp_for_int)[col].agg(
+            {'SRT': fc_to_srt}).reset_index()
+
 
 def srt_dict_to_dataframe(d):
     df_srts = pd.DataFrame()